cleaning up val references [force build]

+++ api/cas-server-core-api-configuration-model/src/main/java/org/apereo/cas/configuration/model/support/consent/ConsentProperties.java
@@ -29,7 +29,7 @@ public class ConsentProperties implements Serializable {
      * Global reminder time unit, to reconfirm consent
      * in cases no changes are detected.
      */
-    private int reminder = 30;
+    private long reminder = 30;
 
     /**
      * Global reminder time unit of measure, to reconfirm consent
