+++ api/cas-server-core-api-configuration-model/src/main/java/org/apereo/cas/configuration/model/support/jdbc/QueryEncodeJdbcAuthenticationProperties.java
@@ -75,7 +75,7 @@ public class QueryEncodeJdbcAuthenticationProperties extends AbstractJpaProperti
     /**
      * Default number of iterations for hashing.
      */
-    private long numberOfIterations;
+    private int numberOfIterations;
 
     /**
      * Static salt to be used for hashing.
